Software/energy-worker/requirements.txt
async-timeout==4.0.2
psycopg2-binary==2.9.5
redis==4.4.0
